+++ documentation/src/pages/reference/types/lucia-types.md
@@ -4,7 +4,33 @@ layout: "@layouts/DocumentLayout.astro"
 title: "Lucia types"
 ---
 
-## LuciaError
+Types can be imported from `lucia-sveltekit/types`.
+
+```ts
+import type { Adapter } from "lucia-sveltekit/types";
+```
+
+## `Adapter`
+
+Refer to [Adapters](/reference/adapters/adapters) reference.
+
+## `ClearSession`
+
+Type for [`clearSessions()`](/reference/api/locals-api#clearsession) locals method.
+
+```ts
+type ClearSession = () => void;
+```
+
+## `GetSession`
+
+Type for [`getSession()`](/reference/api/locals-api#getsession) locals method.
+
+```ts
+type GetSession = () => Session | null;
+```
+
+## `LuciaError`
 
 All errors thrown by Lucia will use this error constructor. Refer to [Errors](/reference/types/errors) for a list of valid error names.
 
@@ -38,6 +64,22 @@ type Session = {
 | userId  | `string` | User id of the user of the session  |
 | expires | `number` | Unix time of the session expiration |
 
+## `SessionAdapter`
+
+Refer to [Adapters](/reference/adapters/adapters) reference.
+
+## `SessionSchema`
+
+Refer to [Database model](/reference/adapters/database-model#schema-type-1) reference.
+
+## `SetSession`
+
+Type for [`setSession()`](/reference/api/locals-api#setsession) locals method.
+
+```ts
+type SetSession = (session: Session) => void;
+```
+
 ## `User`
 
 Return type of [`transformUserData()`](/reference/configure/lucia-configurations#transformuserdata) config.
@@ -56,6 +98,10 @@ type User = {
 };
 ```
 
+## `UserAdapter`
+
+Refer to [Adapters](/reference/adapters/adapters) reference.
+
 ## `UserData`
 
 The columns of `user` table excluding `hashed_password` and `provider_id`.
@@ -70,3 +116,8 @@ type UserData = {
 | ---- | ------------------------------------------------------------------------- | ---------------------------------- |
 | id   | `string`                                                                  | User id of the user                |
 |      | [`Lucia.UserAttributes`](/reference/types/lucia-namespace#userattributes) | Additional columns in `user` table |
+
+
+## `UserSchema`
+
+Refer to [Database model](/reference/adapters/database-model#schema-type-1) reference.
