+++ packages/adapter-test/src/tests/index.ts
@@ -1,5 +1,4 @@
-import type { SessionSchema, UserSchema } from "lucia-sveltekit/adapter";
-import type { Adapter } from "lucia-sveltekit/adapter";
+import type { SessionSchema, UserSchema, Adapter } from "lucia-sveltekit/types";
 import { test, end, validate } from "./../test.js";
 import { User } from "./../db.js";
 import { Database } from "../index.js";
