+++ packages/adapter-test/src/tests/session/index.ts
@@ -1,4 +1,4 @@
-import type { SessionAdapter } from "lucia-sveltekit/adapter";
+import type { SessionAdapter } from "lucia-sveltekit/types";
 import { test, end, validate } from "./../../test.js";
 import { User } from "./../../db.js";
 import { Database } from "../../index.js";
