+++ packages/adapter-test/src/tests/userid/index.ts
@@ -1,4 +1,4 @@
-import type { Adapter } from "lucia-sveltekit/adapter";
+import type { Adapter } from "lucia-sveltekit/types";
 import { test, end, validate } from "./../../test.js";
 import { User } from "./../../db.js";
 import { Database } from "./../../index.js";
