+++ packages/lucia-sveltekit/package.json
@@ -1,6 +1,6 @@
 {
     "name": "lucia-sveltekit",
-    "version": "0.14.1",
+    "version": "0.14.2",
     "description": "A simple authentication library for SvelteKit",
     "main": "index.js",
     "types": "index.d.ts",
