+++ packages/lucia-sveltekit/src/adapter/index.ts
@@ -1,62 +1,4 @@
-export type UserSchema = {
-    id: string;
-    hashed_password: string | null;
-    provider_id: string;
-} & Lucia.UserAttributes;
-
-export type UserData = { id: string } & Required<Lucia.UserAttributes>;
-
-export type SessionSchema = {
-    id: string;
-    expires: number;
-    idle_expires: number;
-    user_id: string;
-};
-
-export type Adapter = {
-    getSessionAndUserBySessionId?: (sessionId: string) => Promise<{
-        user: UserSchema;
-        session: SessionSchema;
-    } | null>;
-} & SessionAdapter &
-    UserAdapter;
-
-export type UserAdapter = {
-    getUser: (userId: string) => Promise<UserSchema | null>;
-    getUserByProviderId: (providerId: string) => Promise<UserSchema | null>;
-    setUser: (
-        userId: string | null,
-        data: {
-            providerId: string;
-            hashedPassword: string | null;
-            attributes: Record<string, any>;
-        }
-    ) => Promise<UserSchema>;
-    deleteUser: (userId: string) => Promise<void>;
-    updateUser: (
-        userId: string,
-        data: {
-            providerId?: string | null;
-            hashedPassword?: string | null;
-            attributes?: Record<string, any>;
-        }
-    ) => Promise<UserSchema>;
-};
-
-export type SessionAdapter = {
-    getSession: (sessionId: string) => Promise<SessionSchema | null>;
-    getSessionsByUserId: (userId: string) => Promise<SessionSchema[]>;
-    setSession: (
-        sessionId: string,
-        data: {
-            userId: string;
-            expires: number;
-            idlePeriodExpires: number;
-        }
-    ) => Promise<void>;
-    deleteSession: (...sessionIds: string[]) => Promise<void>;
-    deleteSessionsByUserId: (userId: string) => Promise<void>;
-};
+import { UserSchema } from "../types.js";
 
 export const getUpdateData = (data: {
     providerId?: string | null;
